Refactor guide events and column placement handling
