+++ packages/core/src/events/bus.ts
@@ -7,7 +7,6 @@ import type {
   ColumnNode,
   DoorNode,
   FenceNode,
-  GuideNode,
   ItemNode,
   LevelNode,
   RoofNode,
@@ -133,12 +132,6 @@ type ToolEvents = {
   'tool:cancel': undefined
 }
 
-type GuideEvents = {
-  'guide:set-reference-scale': { guideId: GuideNode['id'] }
-  'guide:cancel-reference-scale': undefined
-  'guide:deleted': { guideId: GuideNode['id'] }
-}
-
 type PresetEvents = {
   'preset:generate-thumbnail': { presetId: string; nodeId: string }
   'preset:thumbnail-updated': { presetId: string; thumbnailUrl: string }
@@ -180,7 +173,6 @@ type EditorEvents = GridEvents &
   NodeEvents<'door', DoorEvent> &
   CameraControlEvents &
   ToolEvents &
-  GuideEvents &
   PresetEvents &
   ThumbnailEvents &
   SnapshotEvents &
